pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "diffalign"
version = "1.0.0"
description = "DiffAlign: Graph Diffusion Model for Retrosynthesis"
authors = [{name = "N.L", email = "najwa.laabid@aalto.fi"}]
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "networkx",
    "pandas",
    "pillow",
    "rdkit",
    "omegaconf",
    "hydra-core",
    "pyyaml",
    "wandb",
    "pytorch-lightning",
    "matplotlib",
    "seaborn",
    "imageio",
    "fcd_torch",
    "tqdm",
]

[project.optional-dependencies]
api = ["flask", "gunicorn", "markupsafe", "syntheseus", "rxn-insight", "pubchempy"]

[tool.setuptools.packages.find]
include = ["diffalign*", "api*"]
